anova.pima

This is the anova method for pima object. Remark: it performs type III deviance decomposition as in car::Anova.

Usage

# S3 method for class 'pima'
anova(object, ...)

Arguments

object

(the object) glm (or pima) object with the model under the null hypothesis (i.e. the covariates, the nuisance parameters).

...

other parameters allowed in stats::anova.

model1

a glm (or pima) or a matrix (or vector). If it is a glm object, it has the model under the alternative hypothesis. The variables in model1 are the same variables in object plus one or more variables to be tested. Alternatively, if model1 is a matrix, it contains the tested variables column-wise.

score_type

The type of score that is computed. It can be "orthogonalized", "effective" or "basic". Default is "orthogonalized". "effective" and "orthogonalized" take into account the nuisance estimation. The default is NULL, in this case the value is taken from object.

n_flips

The number of random flips of the score contributions. When n_flips is equal or larger than the maximum number of possible flips (i.e. n^2), all possible flips are performed. Default is 5000.

id

a vector identifying the clustered observations. If NULL (default) observations are assumed to be independent. NOTE: if object is a pima and model$flip_param_call$id is not NULL, this is considered in the inference.

Examples

set.seed(1)