Pima Indians diabetes data

A data set of diabetes outcomes and clinical predictors for women from the Pima Indian population.

Usage

data(pimads)

Format

A data frame with 768 rows and 8 variables:

npreg

Number of pregnancies.

glucose

Plasma glucose concentration from an oral glucose tolerance test.

pressure

Diastolic blood pressure in mm Hg.

skin

Triceps skin fold thickness in mm.

bmi

Body mass index.

pedigree

Diabetes pedigree function.

age

Age in years.

diabetes

Diabetes diagnosis, with levels neg and pos.

References

Smith, J. W., Everhart, J. E., Dickson, W. C., Knowler, W. C., and Johannes, R. S. (1988). Using the ADAP learning algorithm to forecast the onset of diabetes mellitus. In Proceedings of the Symposium on Computer Applications in Medical Care.